+++ client/src/assets/sampleData.json
@@ -3,48 +3,100 @@
         "id": 1,
         "image_url": "https://cdn.shopify.com/s/files/1/0317/9853/files/inspiring-christmas-tree-ideas-102538741_large.jpg?v=1481831458",
         "caption": "A christmas tree",
-        "class_categories": {
-            "tree": 0.9,
-            "jewelry": 0.1
-        }
+        "class_categories": [
+            {
+                "className": "cucumber, cuke",
+                "probability": 0.4781543016433716
+            },
+            {
+                "className": "zucchini, courgette",
+                "probability": 0.09959135204553604
+            },
+            {
+                "className": "lemon",
+                "probability": 0.09038704633712769
+            }
+        ],
+        "created_at": "2023-12-19T21:27:50.470Z"
     },
     {
         "id": 2,
         "image_url": "https://hips.hearstapps.com/hmg-prod/images/happy-dog-outdoors-royalty-free-image-1652927740.jpg?crop=0.447xw:1.00xh;0.187xw,0&resize=980:*",
         "caption": "A happy dog",
-        "class_categories": {
-            "dog": 0.8,
-            "cat": 0.2
-        }
+        "class_categories": [
+            {
+                "className": "cucumber, cuke",
+                "probability": 0.4781543016433716
+            },
+            {
+                "className": "zucchini, courgette",
+                "probability": 0.09959135204553604
+            },
+            {
+                "className": "lemon",
+                "probability": 0.09038704633712769
+            }
+        ],
+        "created_at": "2023-12-19T21:27:50.470Z"
     },
     {
         "id": 3,
         "image_url": "https://upload.wikimedia.org/wikipedia/commons/thumb/4/4d/Cat_November_2010-1a.jpg/1200px-Cat_November_2010-1a.jpg",
         "caption": "A grumpy cat.",
-        "class_categories": {
-            "mountain": 0.7,
-            "sky": 0.2,
-            "forest": 0.1
-        }
+        "class_categories": [
+            {
+                "className": "cucumber, cuke",
+                "probability": 0.4781543016433716
+            },
+            {
+                "className": "zucchini, courgette",
+                "probability": 0.09959135204553604
+            },
+            {
+                "className": "lemon",
+                "probability": 0.09038704633712769
+            }
+        ],
+        "created_at": "2023-12-19T21:27:50.470Z"
     },
     {
         "id": 4,
         "image_url": "https://upload.wikimedia.org/wikipedia/commons/thumb/4/4d/Cat_November_2010-1a.jpg/1200px-Cat_November_2010-1a.jpg",
         "caption": "A grumpy cat.",
-        "class_categories": {
-            "mountain": 0.7,
-            "sky": 0.2,
-            "forest": 0.1
-        }
+        "class_categories": [
+            {
+                "className": "cucumber, cuke",
+                "probability": 0.4781543016433716
+            },
+            {
+                "className": "zucchini, courgette",
+                "probability": 0.09959135204553604
+            },
+            {
+                "className": "lemon",
+                "probability": 0.09038704633712769
+            }
+        ],
+        "created_at": "2023-12-19T21:27:50.470Z"
     },
     {
         "id": 5,
         "image_url": "https://upload.wikimedia.org/wikipedia/commons/thumb/4/4d/Cat_November_2010-1a.jpg/1200px-Cat_November_2010-1a.jpg",
         "caption": "A grumpy cat.",
-        "class_categories": {
-            "mountain": 0.7,
-            "sky": 0.2,
-            "forest": 0.1
-        }
+        "class_categories": [
+            {
+                "className": "cucumber, cuke",
+                "probability": 0.4781543016433716
+            },
+            {
+                "className": "zucchini, courgette",
+                "probability": 0.09959135204553604
+            },
+            {
+                "className": "lemon",
+                "probability": 0.09038704633712769
+            }
+        ],
+        "created_at": "2023-12-19T21:27:50.470Z"
     }
 ]
