--- client/src/assets/sampleinsert.json
@@ -1,39 +1,0 @@
-{
-    "caption": "lemons",
-    "class_catgories": [
-        {
-            "className": "cucumber, cuke",
-            "probability": 0.4781543016433716
-        },
-        {
-            "className": "zucchini, courgette",
-            "probability": 0.09959135204553604
-        },
-        {
-            "className": "lemon",
-            "probability": 0.09038704633712769
-        }
-    ],
-    "image_url": "https://plus.unsplash.com/premium_photo-1702056253394-4d1a360e0c15?w=500&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxlZGl0b3JpYWwtZmVlZHwxMzR8fHxlbnwwfHx8fHw%3D"
-}
-
-{
-    "id": 2,
-    "image_url": "https://hips.hearstapps.com/hmg-prod/images/happy-dog-outdoors-royalty-free-image-1652927740.jpg?crop=0.447xw:1.00xh;0.187xw,0&resize=980:*",
-    "caption": "A happy dog",
-    "class_categories": [
-        {
-            "className": "cucumber, cuke",
-            "probability": 0.4781543016433716
-        },
-        {
-            "className": "zucchini, courgette",
-            "probability": 0.09959135204553604
-        },
-        {
-            "className": "lemon",
-            "probability": 0.09038704633712769
-        }
-    ],
-    "created_at": "2023-12-19T21:27:50.470Z"
-  }
